# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface ContactMessageRequest {
  topic: string;
  message: string;
}

export interface ContactMessageResponse {
  ok: boolean;
  cooldownUntil: string;
}

export interface ContactRateLimitResponse {
  error: string;
  cooldownUntil: string;
}

export interface AdminContactUsToggleRequest {
  disabled: boolean;
}

export interface AdminContactUsToggleResponse {
  ok: boolean;
}